리팩토링: PortResolver 서비스 추출
AppsInTossMenu 분할 리팩토링 (6개 PR 중 3번째).

- Editor/Menu/PortResolver.cs 신설 (internal static, namespace AppsInToss.Editor.Menu)
- 6개 메서드 이동: KillProcessOnPort, IsPortAvailable, FindAvailablePort,
  WaitForPortAndOpenBrowser, TryResolveServerPorts, IsPortConflictError
- facade 호출부를 PortResolver.X(...) 로 치환, using 추가
- PortResolverParserTests 추가 (IsPortConflictError 파서 케이스 6건)
