+++ Editor/Menu/PortResolver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7430e0b3f8a24a20b0648079b00399cd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
