+++ Tests~/E2E/SharedScripts/Editor/EditModeTests/Menu/PortResolverParserTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4810930385ae4ad7940987313852368f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
